fix: Added Async task to create and remove tilesets, fixed respones for delete requests
