+++ terraso_backend/apps/graphql/schema/visualization_config.py
@@ -13,22 +13,20 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see https://www.gnu.org/licenses/.
 
-import csv
-import json
-
 import graphene
-import openpyxl
 import structlog
-from django.conf import settings
 from graphene import relay
 from graphene_django import DjangoObjectType
 
-from apps.core.gis.mapbox import create_tileset, get_publish_status, remove_tileset
+from apps.core.gis.mapbox import get_publish_status
 from apps.core.models import Group, Membership
 from apps.graphql.exceptions import GraphQLNotAllowedException
 from apps.shared_data.models.data_entries import DataEntry
 from apps.shared_data.models.visualization_config import VisualizationConfig
-from apps.shared_data.services import data_entry_upload_service
+from apps.shared_data.visualization_tileset_tasks import (
+    start_create_mapbox_tileset_task,
+    start_remove_mapbox_tileset_task,
+)
 
 from ..exceptions import GraphQLNotFoundException
 from .commons import BaseDeleteMutation, BaseWriteMutation, TerrasoConnection
@@ -80,94 +78,6 @@ class VisualizationConfigNode(DjangoObjectType):
             self.mapbox_tileset_status = VisualizationConfig.MAPBOX_TILESET_READY
             self.save()
             return self.mapbox_tileset_id
-
-
-def get_rows_from_file(data_entry):
-    type = data_entry.resource_type
-    if type.startswith("csv"):
-        file = data_entry_upload_service.get_file(data_entry.s3_object_name, "rt")
-        reader = csv.reader(file)
-        return [row for row in reader]
-    elif type.startswith("xls"):
-        file = data_entry_upload_service.get_file(data_entry.s3_object_name, "rb")
-        wb = openpyxl.load_workbook(file)
-        ws = wb.active
-        return [[cell.value for cell in row] for row in ws.iter_rows()]
-    else:
-        raise Exception("File type not supported")
-
-
-def create_mapbox_tileset(data_entry, group_entry, visualization):
-    rows = get_rows_from_file(data_entry)
-
-    first_row = rows[0]
-
-    dataset_config = visualization.configuration["datasetConfig"]
-    annotate_config = visualization.configuration["annotateConfig"]
-
-    longitude_column = dataset_config["longitude"]
-    longitude_index = first_row.index(longitude_column)
-
-    latitude_column = dataset_config["latitude"]
-    latitude_index = first_row.index(latitude_column)
-
-    data_points = annotate_config["dataPoints"]
-    data_points_indexes = [
-        {
-            "label": data_point.get("label", data_point["column"]),
-            "index": first_row.index(data_point["column"]),
-        }
-        for data_point in data_points
-    ]
-
-    annotation_title = annotate_config.get("annotationTitle")
-
-    title_index = (
-        first_row.index(annotation_title)
-        if annotation_title and annotation_title in first_row
-        else None
-    )
-
-    features = []
-    for row in rows:
-        fields = [
-            {
-                "label": data_point["label"],
-                "value": row[data_point["index"]],
-            }
-            for data_point in data_points_indexes
-        ]
-
-        properties = {
-            "title": row[title_index] if title_index else None,
-            "fields": json.dumps(fields),
-        }
-
-        try:
-            longitude = float(row[longitude_index])
-            latitude = float(row[latitude_index])
-            feature = {
-                "type": "Feature",
-                "geometry": {
-                    "type": "Point",
-                    "coordinates": [longitude, latitude],
-                },
-                "properties": properties,
-            }
-
-            features.append(feature)
-        except ValueError:
-            continue
-
-    geojson = {
-        "type": "FeatureCollection",
-        "features": features,
-    }
-    title = f"{settings.ENV} - {visualization.title}"[:64]
-    description = f"{settings.ENV} - {group_entry.name} - {visualization.title}"
-    id = str(visualization.id).replace("-", "")
-    tileset_id = create_tileset(id, geojson, title, description)
-    return tileset_id
 
 
 class VisualizationConfigAddMutation(BaseWriteMutation):
@@ -222,28 +132,8 @@ class VisualizationConfigAddMutation(BaseWriteMutation):
         result = super().mutate_and_get_payload(root, info, **kwargs)
 
         # Create mapbox tileset
-        tileset_id = None
-        try:
-            tileset_id = create_mapbox_tileset(data_entry, group_entry, result.visualization_config)
-        except Exception as error:
-            logger.error(
-                "Error creating mapbox tileset",
-                extra={"data_entry_id": kwargs["data_entry_id"], "error": str(error)},
-            )
+        start_create_mapbox_tileset_task(result.visualization_config.id)
 
-        if tileset_id is not None:
-            result.visualization_config.mapbox_tileset_id = tileset_id
-            try:
-                result.visualization_config.save()
-            except Exception as error:
-                logger.error(
-                    "Error saving mapbox tileset id",
-                    extra={
-                        "data_entry_id": kwargs["data_entry_id"],
-                        "tileset_id": tileset_id,
-                        "error": str(error),
-                    },
-                )
         return cls(visualization_config=result.visualization_config)
 
 
@@ -270,28 +160,12 @@ class VisualizationConfigUpdateMutation(BaseWriteMutation):
                 model_name=VisualizationConfig.__name__, operation=MutationTypes.UPDATE
             )
 
-        # Delete mapbox tileset
-        try:
-            if visualization_config.mapbox_tileset_id:
-                remove_tileset(visualization_config.mapbox_tileset_id)
-        except Exception as error:
-            logger.error(
-                "Error deleting mapbox tileset",
-                extra={"data_entry_id": kwargs["data_entry_id"], "error": str(error)},
-            )
+        result = super().mutate_and_get_payload(root, info, **kwargs)
 
         # Create mapbox tileset
-        try:
-            data_entry = visualization_config.data_entry
-            group_entry = visualization_config.group
-            tileset_id = create_mapbox_tileset(data_entry, group_entry, kwargs)
-            kwargs["mapbox_tileset_id"] = tileset_id
-        except Exception as error:
-            logger.error(
-                "Error creating mapbox tileset",
-                extra={"data_entry_id": kwargs["data_entry_id"], "error": str(error)},
-            )
-        return super().mutate_and_get_payload(root, info, **kwargs)
+        start_create_mapbox_tileset_task(result.visualization_config.id)
+
+        return cls(visualization_config=result.visualization_config)
 
 
 class VisualizationConfigDeleteMutation(BaseDeleteMutation):
@@ -317,13 +191,6 @@ class VisualizationConfigDeleteMutation(BaseDeleteMutation):
             )
 
         # Delete mapbox tileset
-        try:
-            if visualization_config.mapbox_tileset_id:
-                remove_tileset(visualization_config.mapbox_tileset_id)
-        except Exception as error:
-            logger.error(
-                "Error deleting mapbox tileset",
-                extra={"data_entry_id": kwargs["data_entry_id"], "error": str(error)},
-            )
+        start_remove_mapbox_tileset_task(visualization_config.mapbox_tileset_id)
 
         return super().mutate_and_get_payload(root, info, **kwargs)
